requirements.txt
asgiref==3.4.1
autopep8==1.5.5
django-environ==0.4.5
Django==4.0.6
django-cors-headers==3.13.0
django-leaflet==0.28.3
djangorestframework==3.13.1
djangorestframework-gis==1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2


